requirements.txt
django
django-chatbot
wikipedia
django-rest-auth-bearer
